prometheus-exporter/service-code/pip-requirements.txt
Click==7.0
decorator==4.4.1
Flask==1.1.1
gpg==1.10.0
iniparse==0.4
itsdangerous==1.1.0
Jinja2==2.10.3
jsonpatch==1.24
jsonpath-rw==1.4.0
jsonpointer==2.0
MarkupSafe==1.1.1
ply==3.11
prometheus-client==0.7.1
python-dateutil==2.6.1
PyYAML==5.2
recordtype==1.3
redfish==2.1.3
rpm==4.14.2
six==1.11.0
urlparse2==1.1.1
uWSGI==2.0.18
Werkzeug==0.16.0
